# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: gym-logs
    description: Gym check-in log operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /gym-logs:
    post:
      operationId: createGymLog
      tags: [gym-logs]
      summary: Log today's gym check-in
      description: Records whether the authenticated user went to or skipped the gym today. Upserts on the current date.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGymLogBody"
      responses:
        "201":
          description: Log created or updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GymLog"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /gym-logs/stats:
    get:
      operationId: getGymStats
      tags: [gym-logs]
      summary: Get weekly gym stats for the current user
      description: Returns weekly goal, days completed this week, and current streak.
      responses:
        "200":
          description: Stats retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GymStats"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    CreateGymLogBody:
      type: object
      properties:
        status:
          type: string
          enum: [went, skipped]
      required:
        - status
    GymLog:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        date:
          type: string
        status:
          type: string
          enum: [went, skipped]
        createdAt:
          type: string
      required:
        - id
        - userId
        - date
        - status
        - createdAt
    GymStats:
      type: object
      properties:
        weeklyGoal:
          type: integer
        completed:
          type: integer
        streak:
          type: integer
        todayStatus:
          type: ["string", "null"]
          enum: [went, skipped, null]
      required:
        - weeklyGoal
        - completed
        - streak
        - todayStatus
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error