# lib/api-zod/src/generated/types/createGymLogBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateGymLogBodyStatus } from "./createGymLogBodyStatus";

export interface CreateGymLogBody {
  status: CreateGymLogBodyStatus;
}