# lib/api-zod/src/generated/types/gymLog.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GymLogStatus } from "./gymLogStatus";

export interface GymLog {
  id: number;
  userId: string;
  date: string;
  status: GymLogStatus;
  createdAt: string;
}