# lib/api-zod/src/generated/types/gymStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GymStatsTodayStatus } from "./gymStatsTodayStatus";

export interface GymStats {
  weeklyGoal: number;
  completed: number;
  streak: number;
  /** @nullable */
  todayStatus: GymStatsTodayStatus;
}